refactor!: function registration and inner workings
